+++ packages/nimble-components/src/anchor/styles.ts
@@ -1,7 +1,8 @@
 import { css } from '@microsoft/fast-element';
-import { display } from '@microsoft/fast-foundation';
+import { display } from '../utilities/style/display';
 import { focusVisible } from '../utilities/style/focus';
 import {
+    borderHoverColor,
     linkActiveFontColor,
     linkActiveProminentFontColor,
     linkDisabledFontColor,
@@ -62,6 +63,7 @@ export const styles = css`
             outline: none;
             box-shadow: inset 0px -1px;
             text-decoration: underline;
+            color: ${borderHoverColor};
         }
     }