+++ packages/storybook/src/nimble/text-area/text-area.react.tsx
@@ -1,4 +1,4 @@
-import { TextArea } from '@ni/nimble-components/dist/esm/text-area';
+import { TextArea } from '../../../../nimble-components/src/text-area';
 import { wrap } from '../../utilities/react-wrapper';
 
 export const NimbleTextArea = wrap(TextArea);